identitydbcontext ve domain ilişkileri örnekleri eklendi.

+++ etrade-core.domain/Entities/Core/Product.cs
@@ -2,7 +2,7 @@ using etrade_core.domain.Entities.Base;
 
 namespace etrade_core.domain.Entities.Core
 {
-    public class Product : BaseEntity<int>
+    public class Product : BaseEntity<long>
     {
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
