+++ etrade-core.persistence/Repositories/Interfaces/IProductRepository.cs
@@ -2,7 +2,7 @@ using etrade_core.domain.Entities.Core;
 
 namespace etrade_core.persistence.Repositories.Interfaces
 {
-    public interface IProductRepository : IRepository<Product, int>
+    public interface IProductRepository : IRepository<Product, long>
     {
         Task<IEnumerable<Product>> GetActiveProductsAsync();
         Task<IEnumerable<Product>> GetProductsByPriceRangeAsync(decimal minPrice, decimal maxPrice);
